requirements.txt
# for the main streamlit app
streamlit==1.41.1
streamlit-extras==0.5.0
requests==2.32.3
pandas==2.2.3
pydantic==2.10.3
asyncstdlib==3.13.0

# plots
plotly==5.24.1
statsmodels==0.14.4
seaborn==0.13.2

# for the db and django
Django==5.1.4
django-extensions==3.2.3
django-simple-history==3.7.0
djangorestframework==3.15.2
whitenoise==6.8.2
django-colorfield==0.11.0
pretty-html-table==0.9.16
anthropic==0.41.0

# discord bot
discord.py==2.4.0
# easyocr==1.7.2

# fish bot
discord.py==2.4.0
easyocr==1.7.2
pympler==0.9
psutil==5.9.0

# testing
pytest==8.3.4
pytest-django==4.9.0
pytest-sugar==1.0.0

# stuff
tqdm==4.67.1
natsort==8.4.0
schedule==1.2.2
